+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* `OnlyBotThree.dispose()` method to dispose of the bot's underlying three.js resources.
 
 ## [1.3.1](https://github.com/animavirtuality/onlybots-three/compare/1.3.0...1.3.1) - October 7, 2022
 ### Updated